Выполнил задание "Модуляция"
